efi: Fix SNP test log
This commit fixes interface log for snp test.


Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/96465
Tested-by: Suneel Garapati <sgarapati@caviumnetworks.com>

+++ lib/efi_selftest/efi_selftest_snp.c
@@ -370,7 +370,7 @@ static int execute(void)
 			continue;
 		}
 
-		efi_st_printf("Testing eth%ld\n", i);
+		efi_st_printf("Testing eth%d\n", (unsigned int)i);
 		/*
 		 * Check hardware address size.
 		 */
